Models/requirements.txt
opencv-python
mediapipe
numpy
Pillow
tqdm
onnx
onnxscript
onnxruntime